+++ src/frutas/Fruta.java
@@ -3,7 +3,7 @@ package frutas;
 import java.awt.Color;
 import java.awt.Graphics;
 
-import snake.Punto2D;
+import logica.Punto2D;
 
 public class Fruta {
 	Punto2D pos;
