+++ src/gui/AL.java
@@ -1,7 +1,9 @@
-package snake;
+package gui;
 
 import java.awt.event.KeyAdapter;
 import java.awt.event.KeyEvent;
+
+import snake.Orientacion;
 
 
 	public class AL extends KeyAdapter {
