+++ src/gui/EntitySnakeYFruta.java
@@ -1,4 +1,4 @@
-package snake;
+package gui;
 
 import java.awt.Graphics2D;
 import java.awt.Rectangle;
