+++ src/gui/Fondo.java
@@ -1,4 +1,4 @@
-package snake;
+package gui;
 
 import java.awt.Graphics;
 import java.awt.Image;
