+++ src/obstaculo/obstaculoDibujable.java
@@ -3,7 +3,7 @@ package obstaculo;
 import java.awt.Color;
 import java.awt.Graphics;
 
-import snake.Dibujable;
+import logica.Dibujable;
 
 public class obstaculoDibujable extends Dibujable {
