+++ src/snake/Serpiente.java
@@ -5,6 +5,7 @@ import java.util.ArrayList;
 import java.util.Iterator;
 
 import frutas.Fruta;
+import logica.Punto2D;
